Externalize NounTile to own widget.
